Create puma.rb
To set the host and port in Puma web server in Ruby, you can use the -b and -p options followed by the hostname and port number respectively.

Here's an example of how to set the host to localhost and port to 3000:

puma -b tcp://localhost:3000
--------------------------------------
You can also set the host and port in a Puma configuration file. Here's an example of how to set the host to localhost and port to 3000 in a config/puma.rb file:

bind 'tcp://localhost:3000'

You can customize the host and port as per your requirements in the above examples.